Moved classes / interfaces into their own files.
